+++ dice.h
@@ -1,6 +1,6 @@
 #ifndef _DICE_H_
 #define _DICE_H_
 
-bool pixel_is_on(int number, int px);
+bool pixel_is_on(uint8_t number, uint8_t px);
 
 #endif
